# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Zawadi — Premium Kenyan Corporate Gifting Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with optional filters
      parameters:
        - name: category_id
          in: query
          schema: { type: string }
        - name: supplier_id
          in: query
          schema: { type: string }
        - name: occasion
          in: query
          schema: { type: string }
        - name: min_price
          in: query
          schema: { type: number }
        - name: max_price
          in: query
          schema: { type: number }
        - name: search
          in: query
          schema: { type: string }
        - name: is_featured
          in: query
          schema: { type: boolean }
        - name: limit
          in: query
          schema: { type: integer, default: 24 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: Paginated product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a new product (admin/supplier)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a single product
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductInput"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted

  # ── Suppliers ─────────────────────────────────────────────────────────────
...
  /quotes:
    get:
      operationId: listQuotes
      tags: [quotes]
      summary: List quotes for a corporate
      parameters:
        - name: corporate_id
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
      responses:
        "200":
          description: List of quotes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quote"
    post:
      operationId: createQuote
      tags: [quotes]
      summary: Generate an instant quote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuoteInput"
      responses:
        "201":
          description: Generated quote with line items and totals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
  /quotes/{id}:
    get:
      operationId: getQuote
      tags: [quotes]
      summary: Get a quote
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Quote details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
  /quotes/{id}/convert:
    post:
      operationId: convertQuoteToOrder
      tags: [quotes]
      summary: Convert an accepted quote into an order
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # ── Orders ────────────────────────────────────────────────────────────────
...
  /recipients:
    get:
      operationId: listRecipients
      tags: [recipients]
      summary: List recipients for an order
      parameters:
        - name: order_id
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: List of recipients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recipient"
    post:
      operationId: createRecipient
      tags: [recipients]
      summary: Add a single recipient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRecipientInput"
      responses:
        "201":
          description: Created recipient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipient"
  /recipients/bulk:
    post:
      operationId: bulkCreateRecipients
      tags: [recipients]
      summary: Bulk upload recipients (CSV payload) with personalisation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkCreateRecipientsInput"
      responses:
        "201":
          description: Bulk upload result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkRecipientsResult"
  /recipients/{id}:
    put:
      operationId: updateRecipient
      tags: [recipients]
      summary: Update recipient personalisation
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRecipientInput"
      responses:
        "200":
          description: Updated recipient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipient"
    delete:
      operationId: deleteRecipient
      tags: [recipients]
      summary: Remove a recipient
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted

  # ── Invoices ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        details:
          type: string

    # ── Category ──────────────────────────────────────────────────────────
    Category:
      type: object
      required: [id, name, slug]
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        icon:
          type: string
          nullable: true
        product_count:
          type: integer

    # ── Supplier ──────────────────────────────────────────────────────────
    Supplier:
      type: object
      required: [id, name, county, is_verified, onboarding_status, created_at]
      properties:
        id:
          type: string
        name:
          type: string
        contact_name:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        county:
          type: string
        description:
          type: string
          nullable: true
        story:
          type: string
          nullable: true
        cover_image_url:
          type: string
          nullable: true
        logo_url:
          type: string
          nullable: true
        is_verified:
          type: boolean
        onboarding_status:
          type: string
          enum: [pending, under_review, approved, rejected]
        tags:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
...
    Product:
      type: object
      required: [id, name, category_id, supplier_id, unit_price, moq, lead_time_days, origin, is_active, created_at]
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        category_id:
          type: string
        category:
          $ref: "#/components/schemas/Category"
          nullable: true
        supplier_id:
          type: string
        supplier:
          $ref: "#/components/schemas/Supplier"
          nullable: true
        unit_price:
          type: number
          description: Price in KES per unit at MOQ
        bulk_tiers:
          type: array
          items:
            $ref: "#/components/schemas/BulkPriceTier"
        moq:
          type: integer
          description: Minimum order quantity
        lead_time_days:
          type: integer
        origin:
          type: string
          description: County/region of origin
        tags:
          type: array
          items:
            type: string
        occasion_tags:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            $ref: "#/components/schemas/ProductImage"
        is_active:
          type: boolean
        is_featured:
          type: boolean
        weight_grams:
          type: integer
          nullable: true
        dimensions_cm:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    BulkPriceTier:
      type: object
      required: [min_qty, price_per_unit]
      properties:
        min_qty:
          type: integer
        price_per_unit:
          type: number
    ProductImage:
      type: object
      required: [id, url, is_primary]
      properties:
        id:
          type: string
        url:
          type: string
        alt:
          type: string
          nullable: true
        is_primary:
          type: boolean
    CreateProductInput:
      type: object
      required: [name, category_id, supplier_id, unit_price, moq, lead_time_days, origin]
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        category_id:
          type: string
        supplier_id:
          type: string
        unit_price:
          type: number
        bulk_tiers:
          type: array
          items:
            $ref: "#/components/schemas/BulkPriceTier"
        moq:
          type: integer
        lead_time_days:
          type: integer
        origin:
          type: string
        tags:
          type: array
          items:
            type: string
        occasion_tags:
          type: array
          items:
            type: string
        is_featured:
          type: boolean
    UpdateProductInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        unit_price:
          type: number
        bulk_tiers:
          type: array
          items:
            $ref: "#/components/schemas/BulkPriceTier"
        moq:
          type: integer
        lead_time_days:
          type: integer
        tags:
          type: array
          items:
            type: string
        is_active:
          type: boolean
        is_featured:
          type: boolean
    ProductListResponse:
      type: object
      required: [items, total, limit, offset]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer

    # ── Collection ────────────────────────────────────────────────────────
...
    QuoteItem:
      type: object
      required: [product_id, product_name, quantity, unit_price, line_total]
      properties:
        product_id:
          type: string
        product_name:
          type: string
        quantity:
          type: integer
        unit_price:
          type: number
        line_total:
          type: number
        branded_packaging:
          type: boolean
        personalisation:
          type: boolean
    Quote:
      type: object
      required: [id, status, subtotal, vat, total, valid_until, created_at]
      properties:
        id:
          type: string
        corporate_id:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, sent, accepted, rejected, expired]
        items:
          type: array
          items:
            $ref: "#/components/schemas/QuoteItem"
        subtotal:
          type: number
        vat:
          type: number
        total:
          type: number
        notes:
          type: string
          nullable: true
        valid_until:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    CreateQuoteInput:
      type: object
      required: [items]
      properties:
        corporate_id:
          type: string
          nullable: true
        items:
          type: array
          items:
            type: object
            required: [product_id, quantity]
            properties:
              product_id:
                type: string
              quantity:
                type: integer
              branded_packaging:
                type: boolean
              personalisation:
                type: boolean
        notes:
          type: string
          nullable: true

    # ── Order ─────────────────────────────────────────────────────────────
...
    Order:
      type: object
      required: [id, corporate_id, status, subtotal, vat, total, created_at]
      properties:
        id:
          type: string
        reference:
          type: string
          description: Human-readable order reference e.g. ZWD-2026-0042
        corporate_id:
          type: string
        corporate_name:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, confirmed, in_production, quality_check, packaging, dispatched, delivered, cancelled]
        subtotal:
          type: number
        vat:
          type: number
        total:
          type: number
        delivery_address:
          type: string
          nullable: true
        delivery_date:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
        recipient_count:
          type: integer
        created_at:
          type: string
          format: date-time
...
    Recipient:
      type: object
      required: [id, order_id, name]
      properties:
        id:
          type: string
        order_id:
          type: string
        name:
          type: string
        email:
          type: string
          nullable: true
        department:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        personalisation_message:
          type: string
          nullable: true
        delivery_address:
          type: string
          nullable: true
    CreateRecipientInput:
      type: object
      required: [order_id, name]
      properties:
        order_id:
          type: string
        name:
          type: string
        email:
          type: string
          nullable: true
        department:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        personalisation_message:
          type: string
          nullable: true
        delivery_address:
          type: string
          nullable: true
    UpdateRecipientInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          nullable: true
        department:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        personalisation_message:
          type: string
          nullable: true
        delivery_address:
          type: string
          nullable: true
    BulkCreateRecipientsInput:
      type: object
      required: [order_id, recipients]
      properties:
        order_id:
          type: string
        recipients:
          type: array
          items:
            $ref: "#/components/schemas/CreateRecipientInput"
        default_message_template:
          type: string
          nullable: true
          description: "Template with {{name}}, {{title}}, {{department}} placeholders"
    BulkRecipientsResult:
      type: object
      required: [created, failed, total]
      properties:
        created:
          type: integer
        failed:
          type: integer
        total:
          type: integer
        errors:
          type: array
          items:
            type: object
            properties:
              row:
                type: integer
              error:
                type: string

    # ── Invoice ───────────────────────────────────────────────────────────
...