# lib/api-zod/src/generated/types/updateRecipientInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Zawadi — Premium Kenyan Corporate Gifting Platform API
 * OpenAPI spec version: 1.0.0
 */

export interface UpdateRecipientInput {
  name?: string;
  email?: string | null;
  department?: string | null;
  title?: string | null;
  personalisation_message?: string | null;
  delivery_address?: string | null;
}